Dropped `@throws \Exception` report from Repository::sudo
For more details see https://github.com/ezsystems/ezplatform-kernel/pull/339

Requirement of handling \Exception when calling sudo is a bit misleading given a callback can throw any Exception. It should be a responsibility of a developer to know what his callback might throw and either handle it or report on outer method.